+++ template/app/rest/error_handler.go
@@ -33,7 +33,7 @@ func DLiveRHTTPErrorHandler(err error, eCtx echo.Context) {
 		default:
 			//%:{{ `
 			err = skeleton.SemanticError{Err: eErr, Fields: []interface{}{"method", eCtx.Request().Method, "url", eCtx.Request().URL.String()}}.E()
-			//%: ` | replace "skeleton" .SchemaPackage }}
+			//%: ` | replace "skeleton" .SchemaPackage | trim }}
 		}
 	}