+++ template/app/schema/skeleton/http.go
@@ -10,4 +10,11 @@ type DummyRequest struct {
 	DummyParam string    `param:"dummy_param"`
 	DummyGroup string    `json:"dummy_group" validate:"required"`
 	DummyUUID  uuid.UUID `json:"dummy_uuid"`
+}
+
+//%: {{ if and .Yafuds .Examples }}
+type UpdateDummyRequest struct {
+	DummyUUID uuid.UUID `param:"dummy_uuid" validate:"required,uuid"`
+	BaseEditable
 } //%: {{ end }}
+//%: {{ end }}